Demo/requirements.txt
streamlit
joblib
pandas
scikit-learn
numpy
Pillow
